Assignment1: bugfix, reserved spaces in queue
Round-Robin segfaulted because consumers tried to pop a "non-empty" queue, which
was really empty because a consumer was working on the product and had
"reserved" the space. Added a second 'realCount' to the queue for the pop
functions to use.

Also made Round-Robin update the life of products so that it doesn't infinite
loop on the first set of products.