Assignment2: tracking page swaps
The program now has the ability to track page swaps in an unsigned long counter.
The program also prints the number of page swaps at the end of execution and the
`plot.sh` script was updated so that it will get this number and properly setup
it's data files.

Also added Makefile targets to aid testing and creating plots.