Assignment2: only load up to all pages
If a process doesn't have the number of pages that it is allowed to load into
main memory initially that is fine, it just has blank spaces associated with it.
Page table structs now store a `numLoaded` so that it is possible to find how
many pages a program has loaded.