Assignment2: fixed segfault issue
Made sure that the number of pages created is done in the same way that a memory
location is looked up. (both use the `ceil` function now to avoid trying to find
a page that doesn't exist)

Created macros for page replacement methods, this means we only need to change
them in one place if we need to, and insures against typos when comparing in
other places.

Also updated Makefile target tar to create the tarball two levels up, so it's in
the main assignments directory and not the source directory.