Assignment2: bugfixes, fixed page replacements
Added some debug statements a couple error messages and fixed errors.

Bugs existed in popping the FIFOs and getting the index of LRU pages causing -1
to be returned in all cases and no swapping to actually be done. These are now
corrected and all the page replacement methods behave as expected.