Assignment3: starting to implement ldisk
Added functions to create blocks and allocate space in ldisk for files. The
allocate function is not finished. Files now also keep track of their number of
blocks allocated.

Also moved initialization of gl.numBlocks to the `init.c` file as it doesn't
have anything to do with parsing arguments.