Assignment3: properly adding files and directories
Added `findInHierarchy`, `addToHierarchy` and `countPathSeparations` functions
in order to help add files and directories to their proper places in the global
hierarchy tree.

Added a `#define` constant for `PATH_SEP`.

Added `cur_dir` to the global environment so that it can keep track of the
current node in the tree that the user is at.

Added a Makefile target `test` which will compile in debug mode and start the
program.

Note: There is currently a bug somewhere in `allocateFile` which causes a
segfault, this need to be fixed.