Assignment3: fixed bugs <- lol so generic
Added core dumps to the `clean` target of the Makefile and increased disk size
for the `test` target.

In `allocateFile` made sure to wrap the created block in a node before appending
it to the linked_list. Did the same in `init` also.

In `main` zero'd out line so that valgrind can be sure the value is initialized
before it is used.

In `str2vect` made sure to malloc the correct number of bytes.